station names getting + many fixes

+++ app/src/main/java/com/svadev/eca/models/AuthCharacterModel.kt
@@ -10,6 +10,6 @@ class AuthCharacterModel (
     val IntellectualProperty: String?="",
     val access_token: String?="",
     val token_type: String?="",
-    val expires_in: Int=0,
+    val expires_in: Long=0,
     val refresh_token: String?=""
 )
